Split e2e tests by workflow
